updates to vuforia 8.1, including updates vuforia sample app code to fix for new api
slight cleanup to ar code

updates to calculate correct projection matrix (near and far plane were incorrect, and needed to wait until state fully loads camera calibration). NOTE: currently the 5th, 8th, 9th, 10th, and 11th elements of the projection matrix need to be *= -1 in order to match the previous version exactly. we might be able to use the new one directly and remove a lot of unnecessary correctedCamera calculations in the future

adds Vuforia.framework to .gitignore because it is too large to push to git

.gitignore fix

updates gitignore again

update gitignore